requirements.txt
telebot
firebase-admin
firestore
asyncio
python-dotenv
aiohttp